Move alloc tests.
